Added onToggle effect for double clicking a task
Did the prop drilling thing again where we passed a prop onToggle all the way down to the singular task. In app.js, we set the toggleReminder func to set the reminder of the task that was clicked to the opposite. In the Task component, we have some conditional rendering as well as setting the onclick event for the individual task